third_party/blink/web_tests/platform/generic/fast/events/touch/page-scaled-touch-gesture-click-expected.txt

This tests basic single touch gesture generation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS event.type is "mousemove"
PASS event.clientX is 20
PASS event.clientY is 24
PASS event.shiftKey is false
PASS event.altKey is false
PASS event.ctrlKey is false
PASS event.metaKey is false
PASS event.type is "mousedown"
PASS event.clientX is 20
PASS event.clientY is 24
PASS event.shiftKey is false
PASS event.altKey is false
PASS event.ctrlKey is false
PASS event.metaKey is false
PASS event.type is "mouseup"
PASS event.clientX is 20
PASS event.clientY is 24
PASS event.shiftKey is false
PASS event.altKey is false
PASS event.ctrlKey is false
PASS event.metaKey is false
PASS event.type is "click"
PASS event.clientX is 20
PASS event.clientY is 24
PASS event.shiftKey is false
PASS event.altKey is false
PASS event.ctrlKey is false
PASS event.metaKey is false
PASS successfullyParsed is true

TEST COMPLETE
